keyframes-method added.
Make animations with @keyframe like this:
```php5
<?=WS()->keyframes("name".
	WS(25)->color("red")->return,
	WS(50)->background("orange")->return,
	WS(100)->color("white")->return
)?>
```

Also added the ->return method to really only return the data - as
__ONE LINE__. It sets WS()->beauty=false while operating like ->end and
then resets it to what it was before.

Also added __reset() method to reset the singleton to its default
values - does not unload any class.

Re-wrote the addDefs method a little and made it fit with the existing
classes.

…Also a new WIP rule - animation.