Added import-rule, changed WingStyleManager::__call
The __call method now tries to find the rule that oyu are looking for.
If it can't however, it's just adding a "fake"-rule. Like when you call
->height (which I just deleted :)) it will indeed add the height rule.
Be careful, when you add an unsupported rule that way, you might
accidentially produce false CSS. That is why we're prepending a /*WS*/
to these rules to mark them as a result of such a call.
WingStyleDesigner::format is now a public-static method, you can use it
outside of it as well…well it happens inside the code, but maybe you
want to take advantage of that.